Adjust the render order between border and rounded window
if border width is a positive value (border be drawn inside of window),
we will draw rounded corners first, then mix it with border.

if border width is a negative value (border be drawn outside of window),
we will draw border color first, then mix it with rounded window.

relative to #67